alx_travel_app/requirement.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
drf-yasg>=1.21.7
celery>=5.3.4
mysqlclient>=2.2.0
django-environ>=0.11.2